refactor(modals): add onSubmit callback and a common cancel button

+++ packages/client/src/modals/project/draft/index.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 
 import { Button } from '$components/button';
 import { Modal, ModalActions } from '$components/modal';
-import { useModal, useModalContext } from '$hooks/useModal';
+import { useModal } from '$hooks/useModal';
 import { ProjectRename } from '$modals/project/rename';
 import { BaseProject } from '$stores/project';
 
@@ -12,36 +12,28 @@ interface Props {
 }
 
 export const ProjectDraft: React.FC<Props> = observer(({ project }) => {
-  const { close } = useModalContext();
   const { show } = useModal(
     <ProjectRename project={project} closeAfterRenaming />,
   );
 
   return (
-    <Modal title="Draft project" size="small">
+    <Modal
+      title="Draft project"
+      size="small"
+      onSubmit={async (_, action) => {
+        if (action === 'delete') {
+          await project.delete();
+        } else {
+          show();
+        }
+      }}
+    >
       <div>This project is a draft. Would you like to save it?</div>
       <ModalActions>
-        <Button variant="secondary" onClick={() => close()}>
-          Cancel
-        </Button>
-        <Button
-          variant="danger"
-          onClick={() => {
-            project.delete();
-            close();
-          }}
-        >
+        <Button variant="danger" action="delete">
           Delete
         </Button>
-        <Button
-          variant="primary"
-          onClick={() => {
-            close();
-            show();
-          }}
-        >
-          Save
-        </Button>
+        <Button variant="primary">Save</Button>
       </ModalActions>
     </Modal>
   );
